fixed an issue with AR
