+++ src/screens/ProfileScreen.js
@@ -55,6 +55,15 @@ export default class LoginScreen extends Component {
       </View>
     );
   }
+
+   // Returns the ViroARSceneNavigator which will start the AR experience
+   _getARNavigator() {
+    return (
+      <ViroARSceneNavigator {...this.state.sharedProps}
+        initialScene={{scene: InitialARScene}} />
+    );
+  }
+
   _goToEditProfileScreen() {
     return (
       <EditProfileScreen />
